Android: Move events handling to QtLayout
Whenever a widget with Qt::Window flag is used a new
QtSurface is created and this is used by Android to
send us events. However, as soon as this Surface is
destroyed, the events stop coming until the user
starts the new event by touching the new Surface.
This causes interruptions and leaves events in an
undefined state in apps where events are supposed to
come in continuous manner.

By moving it to ViewGroup, we stop relying on Surfaces.

[ChangeLog][Android] Move events handling to QtLayout